Added interfaces for Services and implemented them.Made controller updates.
