+++ src/main/java/com/bancx/assignment/loanpaymentsystem/loan/repository/LoanRepository.java
@@ -2,6 +2,8 @@ package com.bancx.assignment.loanpaymentsystem.loan.repository;
 
 import com.bancx.assignment.loanpaymentsystem.loan.model.Loan;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface LoanRepository extends JpaRepository<Loan, Long> {
 }
