+++ src/main/java/com/bancx/assignment/loanpaymentsystem/payment/controller/PaymentController.java
@@ -2,8 +2,8 @@ package com.bancx.assignment.loanpaymentsystem.payment.controller;
 
 import com.bancx.assignment.loanpaymentsystem.payment.dto.PaymentRequestDto;
 import com.bancx.assignment.loanpaymentsystem.payment.dto.PaymentResponseDto;
-import com.bancx.assignment.loanpaymentsystem.payment.model.Payment;
 import com.bancx.assignment.loanpaymentsystem.payment.service.PaymentService;
+import com.bancx.assignment.loanpaymentsystem.payment.service.impl.PaymentServiceImpl;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -17,8 +17,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/payments")
 public class PaymentController {
 
-    @Autowired
-    private PaymentService paymentService;
+    private final PaymentService paymentService;
+
+    public PaymentController(PaymentService paymentService) {
+        this.paymentService = paymentService;
+    }
 
     @PostMapping
     public ResponseEntity<PaymentResponseDto> makePayment(@Valid @RequestBody PaymentRequestDto dto) {
