+++ src/main/java/com/bancx/assignment/loanpaymentsystem/payment/repository/PaymentRepository.java
@@ -2,5 +2,7 @@ package com.bancx.assignment.loanpaymentsystem.payment.repository;
 
 import com.bancx.assignment.loanpaymentsystem.payment.model.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface PaymentRepository extends JpaRepository<Payment, Long> {}
